[OpenCL] Fixes adam_test fail for SYCL (#21)
* Adds SYCL version of UnsortedSegmentSum

Uses a copy of the CPU version to provide an initial SYCL implementation
of the kernel. Doesn't add the other Segment ops which are included in
this file.

* Adds and registers SYCL version of GatherOp

Needed to change build file to include a dependency for gather_functor
on full tensorflow framework, rather than framework-lite.

The GatherOp uses the GatherFunctor, which is similar to the CPU
version, only the fast memcpy has been removed, so that Eigen handles
everything.

* Registers int64 SYCL version of Reshape op

Needed for the AdamOptimizer.

* Puts UnsortedSegmentSum segment_ids in host memory

The way the current kernel works on SYCL devices requires the
segment_ids to be accessed by the host, and so need to be available in
the host memory.

* Removes unwanted commented out code

* Only require full framework for SYCL, not CPU

The register_types.h header (providing TF_CALL_...) pulls in some
protobuf headers, which are only provided by the full framework, not the
framework_lite. Add this dependency when building with SYCL support,
but not when building for CPU only.

* Fixes spelling