[OpenCL] Fix allocator destruction race condition (#136)
* [OpenCL] Changes SYCL Interface construction

Uses C++11 static initialisation to provide singleton instance, rather
than a mutex and pointer.

* [OpenCL] Adds const to SYCL Interface methods

* [OpenCL] Fix allocator destruction race condition

A Tensor's allocator must outlive it, however there is no easy way to
determine whether an Allocator has any Tensors still alive, and so we
cannot know when it is safe to destroy an allocator. The CPU allocator
gets round this by being deleted, so we adopt this convention here.

* [OpenCL] Reformats SYCL code

* [OpenCL] Fixes SYCL comments

* [OpenCL] Tidies SYCL device description

Adds check for whether the QueueInterface pointer is valid, as this
may not always be the case.

* [OpenCL] Adds nullptr checking to SYCL allocator

* [OpenCL] Adds const specifier to SYCL Interface