[OpenCL] Improves BiasOp and ApplyMomentum performance (#146)
* [OpenCL] Reworks BiasOp for all dimensions

We can flatten and broadcast the tensors in BiasOp, so don't actually
need to use the tensor dimensions. In BiasGradOp instead of creating a
FlatTensor which we then reshape to two dimensions in the Eigen
expression, we can create a 2D tensor to start with and not need the
additional step in the expression.

* [OpenCL] Improves SYCL ApplyMomentum

Eigen has better performance if we can avoid using reshape
for now. By changing the ConstScalar to a TensorMap around a
one dimensional tensor we can skip the reshape.