[OpenCL] Fixes zeroing tensor on wrong device (#159)
Eigen's setZero only runs on the CPU which causes a problem if the
pointer is actually a device pointer. Instead change it to out.device(d)
= out.constant(0), which gives exactly the same result.

Although the comment in the bias_op code suggests that suing anything
other than setZero might cause a crash, this is not the case here.