[SNN] Uses conservative memory limits to prevent crashes
Currently both winograd and im2col convolutions require large
intermediate tensors, which are hard to allocate on the device. We need
to use very conservative allocation limits to ensure that all the
tensors required by tensorflow will be able to be allocated.

The additional synchronization added also helps to ensure that we don't
try to allocate too much at any one time, instead waiting for the device
to finish computations.